basic framework for connection managers, still many tests to clean up [skip ci]
