+++ dbt/adapters/bigquery/__init__.py
@@ -1,2 +1,3 @@
+from dbt.adapters.bigquery.connections import BigQueryConnectionManager
 from dbt.adapters.bigquery.relation import BigQueryRelation
 from dbt.adapters.bigquery.impl import BigQueryAdapter
