+++ dbt/adapters/postgres/__init__.py
@@ -1,1 +1,2 @@
+from dbt.adapters.postgres.connections import PostgresConnectionManager
 from dbt.adapters.postgres.impl import PostgresAdapter
