+++ dbt/adapters/redshift/__init__.py
@@ -1,1 +1,2 @@
+from dbt.adapters.redshift.connections import RedshiftConnectionManager
 from dbt.adapters.redshift.impl import RedshiftAdapter
