+++ dbt/adapters/snowflake/__init__.py
@@ -1,2 +1,3 @@
+from dbt.adapters.snowflake.connections import SnowflakeConnectionManager
 from dbt.adapters.snowflake.relation import SnowflakeRelation
 from dbt.adapters.snowflake.impl import SnowflakeAdapter
