+++ dbt/adapters/sql/__init__.py
@@ -1,1 +1,2 @@
+from dbt.adapters.sql.connections import SQLConnectionManager
 from dbt.adapters.sql.impl import SQLAdapter
